feat: realyze eye for show password
